add_navController in bottomNavigation fix name package
