+++ app/src/main/java/com/ncbs/dictionary/data/WordService.kt
@@ -1,5 +1,6 @@
-package com.ncbs.dictionary.repository
+package com.ncbs.dictionary.data
 
+import com.ncbs.dictionary.data.WordDto
 import retrofit2.http.GET
 
 interface WordService {
