+++ app/src/main/java/com/ncbs/dictionary/domain/Word.kt
@@ -1,5 +1,7 @@
 package com.ncbs.dictionary.domain
 
+import com.ncbs.dictionary.domain.LocaleData
+
 data class Word(
     val id: String,
     val locales: Map<String, LocaleData>,
