+++ app/src/main/java/com/ncbs/dictionary/domain/WordListItem.kt
@@ -1,0 +1,6 @@
+package com.ncbs.dictionary.domain
+
+data class WordListItem(
+    val word: Word,
+    val title: String,
+)
