--- app/src/main/java/com/ncbs/dictionary/presentation/WordListItem.kt
@@ -1,8 +1,0 @@
-package com.ncbs.dictionary.presentation
-
-import com.ncbs.dictionary.domain.Word
-
-data class WordListItem(
-    val word: Word,
-    val title: String,
-)
