+++ app/src/main/java/com/ncbs/dictionary/ui/activity/AboutActivity.kt
@@ -1,4 +1,4 @@
-package com.ncbs.dictionary.presentation
+package com.ncbs.dictionary.ui.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
