--- app/src/main/res/layout/content_main.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/navHostFragmentContentMain"
-    android:layout_width="0dp"
-    android:layout_height="0dp" />
